Updated version of an old Vine-interface patch
This extends the C interface with another counterexample format,
which the OCaml interface code used by Vine (and FuzzBALL, etc.)
uses. I believe that historically the reason for this interface
was to get a counterexample that was more similar to the
counterexample printed on the command line, perhaps in regards
to how unconstrained variables were treated. It would be nice
to figure out if we can remove this special patch, either by
making the Vine code happy with STP as it is now, or understanding
what Vine wants as a feature that could be added to the
upstream STP.