Try to avoid boost::pool's boost_system library use
This is a bit of a hack: there's a #define that sounds like it should
have this effect, but in the version I was looking at (1.55), it appears
you also need to fake an include guard. This is ugly, but so IMO is
requiring every program that links with STP to know about how to
link with the appropriate boost_system library.