Quiet enum/int mixture warning
Might be better style to use a const int for N, or at least g++'s
warnings think so.
The signed/unsigned mixing here is also weird, but I think doing the
comparison as unsigned is safest: neither negative values nor values
so large they can only be represented as unsigned are allowed.