Run "lit" with its own choice of Python interpreter
In my testing this is needed to let a "virtualenv"-based version of
lit work correctly, which is in turn needed when a user doesn't have
permission to install Python modules globally on a system. The change
to tests/query-files/CMakeLists.txt is the one that actually made the
difference in my testing, while the change to lit.site.cfg.in
embodies the same thinking (this file is Python code, so if it's
running, we must have already found a working Python interpreter).